feat: Update URLs to archive-docs.ml5js.org
This commit updates the URLs in the CNAME, README.md, and index.html files to point to the archive-docs.ml5js.org domain instead of learn.ml5js.org. This change ensures that users are directed to the archived documentation site after the migration. The URLs for Getting Started, API Reference, Community, and FAQ have all been updated accordingly. Additionally, the logo image URL in the index.html file has been updated to the new domain. The standalone examples section in the README.md file now includes a link to the examples directory on GitHub instead of the ml5.js examples index website. These changes improve the accuracy and consistency of the URLs in the documentation.